requirements.txt
arxiv
feedparser
jinja2
google-genai
aiohttp
python-dateutil
pytz
requests
python-dotenv
